Move file transfer logic from RequestOld to FileTransferManager.java
